requirements.txt
kiteconnect==4.2.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
python-telegram-bot==20.5
numpy==1.24.4 